Restore PYTHONPATH bootstrap patch.
This patch was previously applied to the debian/rules template for this package to facilitate finding the ROS installation prefix's python libraries without the ros_workspace package since this is it.

The rosdistro migration tools[1] ought to have preserved these patches however I suspect that behavior is not working correctly when the source and destination distribution are the same.

[1]: https://github.com/ros/rosdistro/tree/master/migration-tools